unify createNodesFromDOM
