--- packages/lexical-html/src/$createNodesFromDOM.ts
@@ -1,149 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-import {
-  $createLineBreakNode,
-  $createParagraphNode,
-  $isBlockElementNode,
-  $isElementNode,
-  $isRootOrShadowRoot,
-  ArtificialNode__DO_NOT_USE,
-  type DOMChildConversion,
-  isBlockDomNode,
-  type LexicalEditor,
-  type LexicalNode,
-} from 'lexical';
-
-import {IGNORE_TAGS} from './constants';
-import {getConversionFunction} from './getConversionFunction';
-import {isDomNodeBetweenTwoInlineNodes} from './isDomNodeBetweenTwoInlineNodes';
-import {$wrapContinuousInlines} from './wrapContinuousInlines';
-
-export function $createNodesFromDOM(
-  node: Node,
-  editor: LexicalEditor,
-  allArtificialNodes: Array<ArtificialNode__DO_NOT_USE>,
-  hasBlockAncestorLexicalNode: boolean,
-  forChildMap: Map<string, DOMChildConversion> = new Map(),
-  parentLexicalNode?: LexicalNode | null | undefined,
-): Array<LexicalNode> {
-  let lexicalNodes: Array<LexicalNode> = [];
-
-  if (IGNORE_TAGS.has(node.nodeName)) {
-    return lexicalNodes;
-  }
-
-  let currentLexicalNode = null;
-  const transformFunction = getConversionFunction(node, editor);
-  const transformOutput = transformFunction
-    ? transformFunction(node as HTMLElement)
-    : null;
-  let postTransform = null;
-
-  if (transformOutput !== null) {
-    postTransform = transformOutput.after;
-    const transformNodes = transformOutput.node;
-    currentLexicalNode = Array.isArray(transformNodes)
-      ? transformNodes[transformNodes.length - 1]
-      : transformNodes;
-
-    if (currentLexicalNode !== null) {
-      for (const [, forChildFunction] of forChildMap) {
-        currentLexicalNode = forChildFunction(
-          currentLexicalNode,
-          parentLexicalNode,
-        );
-
-        if (!currentLexicalNode) {
-          break;
-        }
-      }
-
-      if (currentLexicalNode) {
-        lexicalNodes.push(
-          ...(Array.isArray(transformNodes)
-            ? transformNodes
-            : [currentLexicalNode]),
-        );
-      }
-    }
-
-    if (transformOutput.forChild != null) {
-      forChildMap.set(node.nodeName, transformOutput.forChild);
-    }
-  }
-
-  // If the DOM node doesn't have a transformer, we don't know what
-  // to do with it but we still need to process any childNodes.
-  const children = node.childNodes;
-  let childLexicalNodes = [];
-
-  const hasBlockAncestorLexicalNodeForChildren =
-    currentLexicalNode != null && $isRootOrShadowRoot(currentLexicalNode)
-      ? false
-      : (currentLexicalNode != null &&
-          $isBlockElementNode(currentLexicalNode)) ||
-        hasBlockAncestorLexicalNode;
-
-  for (let i = 0; i < children.length; i++) {
-    childLexicalNodes.push(
-      ...$createNodesFromDOM(
-        children[i],
-        editor,
-        allArtificialNodes,
-        hasBlockAncestorLexicalNodeForChildren,
-        new Map(forChildMap),
-        currentLexicalNode,
-      ),
-    );
-  }
-
-  if (postTransform != null) {
-    childLexicalNodes = postTransform(childLexicalNodes);
-  }
-
-  if (isBlockDomNode(node)) {
-    if (!hasBlockAncestorLexicalNodeForChildren) {
-      childLexicalNodes = $wrapContinuousInlines(
-        node,
-        childLexicalNodes,
-        $createParagraphNode,
-      );
-    } else {
-      childLexicalNodes = $wrapContinuousInlines(
-        node,
-        childLexicalNodes,
-        () => {
-          const artificialNode = new ArtificialNode__DO_NOT_USE();
-          allArtificialNodes.push(artificialNode);
-          return artificialNode;
-        },
-      );
-    }
-  }
-
-  if (currentLexicalNode == null) {
-    if (childLexicalNodes.length > 0) {
-      // If it hasn't been converted to a LexicalNode, we hoist its children
-      // up to the same level as it.
-      lexicalNodes = lexicalNodes.concat(childLexicalNodes);
-    } else {
-      if (isBlockDomNode(node) && isDomNodeBetweenTwoInlineNodes(node)) {
-        // Empty block dom node that hasnt been converted, we replace it with a linebreak if its between inline nodes
-        lexicalNodes = lexicalNodes.concat($createLineBreakNode());
-      }
-    }
-  } else {
-    if ($isElementNode(currentLexicalNode)) {
-      // If the current node is a ElementNode after conversion,
-      // we can append all the children to it.
-      currentLexicalNode.append(...childLexicalNodes);
-    }
-  }
-
-  return lexicalNodes;
-}
